+++ discovery-provider/src/api/v1/helpers.py
@@ -98,6 +98,27 @@ def add_user_artwork(user):
     return user
 
 
+# Helpers
+def extend_search(resp):
+    if "users" in resp:
+        resp["users"] = list(map(extend_user, resp["users"]))
+    if "followed_users" in resp:
+        resp["followed_users"] = list(map(extend_user, resp["followed_users"]))
+    if "tracks" in resp:
+        resp["tracks"] = list(map(extend_track, resp["tracks"]))
+    if "saved_tracks" in resp:
+        resp["saved_tracks"] = list(map(extend_track, resp["saved_tracks"]))
+    if "playlists" in resp:
+        resp["playlists"] = list(map(extend_playlist, resp["playlists"]))
+    if "saved_playlists" in resp:
+        resp["saved_playlists"] = list(map(extend_playlist, resp["saved_playlists"]))
+    if "albums" in resp:
+        resp["albums"] = list(map(extend_playlist, resp["albums"]))
+    if "saved_albums" in resp:
+        resp["saved_albums"] = list(map(extend_playlist, resp["saved_albums"]))
+    return resp
+
+
 def extend_user(user, current_user_id=None):
     user_id = encode_int_id(user["user_id"])
     user["id"] = user_id
