+++ discovery-provider/src/queries/search_queries.py
@@ -7,7 +7,7 @@ from functools import cmp_to_key
 import sqlalchemy
 from flask import Blueprint, request
 from src import api_helpers, exceptions
-from src.api.v1.search import extend_search
+from src.api.v1.helpers import extend_search
 from src.models import Follow, RepostType, Save, SaveType
 from src.queries import response_name_constants
 from src.queries.get_unpopulated_playlists import get_unpopulated_playlists
